+++ gulpfile.js
@@ -15,7 +15,6 @@ gulp.task('build:css-json', function () {
             strip: /.css/
         }))
         .pipe(gulp.dest('./.trash'));
-    // del('trash/**');
 });
 var del = require('del');
 gulp.task('garbage-collect', function () {
@@ -61,8 +60,10 @@ var concat = require('gulp-concat');
 var uglify = require('gulp-uglify');
 gulp.task('build:js', function () {
     return gulp.src([
+
         // CodeMirror Core
         codemirrorPath + 'lib/codemirror.js',
+
         // Modes
         codemirrorPath + 'mode/xml/xml.js',
         codemirrorPath + 'mode/javascript/javascript.js',
@@ -72,8 +73,33 @@ gulp.task('build:js', function () {
         codemirrorPath + 'mode/php/php.js',
         './src/shortcode.js',
         './src/wordpresspost.js',
+
         // AddOns
         codemirrorPath + 'addon/selection/active-line.js',
+        codemirrorPath + 'addon/search/searchcursor.js',
+        codemirrorPath + 'addon/search/search.js',
+        codemirrorPath + 'addon/dialog/dialog.js',
+        codemirrorPath + 'addon/scroll/simplescrollbars.js',
+        codemirrorPath + 'addon/comment/comment.js',
+        
+        codemirrorPath + 'addon/fold/foldcode.js',
+        codemirrorPath + 'addon/fold/foldgutter.js',
+        codemirrorPath + 'addon/fold/xml-fold.js',
+        // codemirrorPath + 'addon/fold/brace-fold.js', // for JS
+        // codemirrorPath + 'addon/fold/comment-fold.js',
+        codemirrorPath + 'addon/fold/indent-fold.js',
+        
+        codemirrorPath + 'addon/edit/matchbrackets.js',
+        codemirrorPath + 'addon/edit/matchtags.js',
+        codemirrorPath + 'addon/search/match-highlighter.js',  
+        codemirrorPath + 'addon/edit/closetag.js',
+        codemirrorPath + 'addon/edit/closebrackets.js',
+
+        codemirrorPath + 'keymap/sublime.js',
+        codemirrorPath + 'keymap/emacs.js',
+        codemirrorPath + 'keymap/vim.js',
+
+        // ... and finally ...
         // HESH
         './src/hesh.js'
     ])