+++ html-editor-syntax-highlighter.php
@@ -2,7 +2,7 @@
 
 
 /**
- * @since              2.0.0
+ * @since              2.2.0
  * @package            HESH_plugin
  *
  * Plugin Name:        HTML Editor Syntax Highlighter - DEV
@@ -17,11 +17,11 @@
  * License URI:        http://www.gnu.org/licenses/gpl-2.0.txt
  * GitHub Branch:      master
  * GitHub Plugin URI:  https://github.com/mukhortov/HESH-WordPress-Plugin
- * Version:            2.0.0
+ * Version:            2.2.0
  * Requires at least:  4.0.15
- * Tested up to:       4.7.3
- * Stable tag:         2.0.0
- **/
+ * Tested up to:       4.8.1
+ * Stable tag:         2.2.0
+**/
 
 // Check for required PHP version
 if ( version_compare( PHP_VERSION, '5.2.17', '<' ) ) {
@@ -84,17 +84,17 @@ class wp_html_editor_syntax {
 	private $nonceSecretCode = 'secret-code';
 	private $prefix = 'hesh_';
 	private $userPrefrences; // added to the primary bar
-	private $addOns; // added to the advanced menu
 	public function hesh_set_options() {
 		$this->userPrefrences = array(
+
+			// PRIMARY OPTIONS // added in the primary settings bar
 			'theme' => array(
 				'title' => 'Theme',
 				'type' => 'select',
-				// http://stackoverflow.com/questions/18881693/how-to-import-external-json-and-display-in-php
-				// TODO: add this in without a file_get_contents()
-				'options' => json_decode(file_get_contents(dirname(__FILE__) . '/css.json'), true), 
+				'options' => $this->cssThemes,
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'theme' , true),
 				'default' => 'material',
+				'set' => 'primary',
 			),
 			'tabSize' => array(
 				'title' => 'Indent',
@@ -102,20 +102,21 @@ class wp_html_editor_syntax {
 				'options' => range(1,6),
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'tabSize' , true),
 				'default' => 4,
+				'set' => 'primary',
 			),
 			'lineWrapping'=> array(
 				'title' => 'Line Wrap',
 				'type' => 'checkbox',
-				'text' => 'Wrap lines',
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'lineWrapping' , true),
 				'default' => true,
+				'set' => 'primary',
 			),
 			'lineNumbers'=> array(
 				'title' => 'Numbering',
 				'type' => 'checkbox',
-				'text' => 'Show line numbers',
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'lineNumbers' , true),
 				'default' => true,
+				'set' => 'primary',
 			),
 			'fontSize'=> array(
 				'title' => 'Font Size',
@@ -123,6 +124,7 @@ class wp_html_editor_syntax {
 				'options' => range(8,20),
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'fontSize' , true),
 				'default' => 13,
+				'set' => 'primary',
 			),
 			'lineHeight'=> array(
 				'title' => 'Line Height',
@@ -130,34 +132,81 @@ class wp_html_editor_syntax {
 				'options' => range(1,2,0.25),
 				'current' => get_user_meta( get_current_user_id(), $this->prefix.'lineHeight' , true),
 				'default' => 1.5,
+				'set' => 'primary',
+			),
+
+			// ADVANCED OPTIONS // added to the advanced dropdown
+			'matchBrackets'=> array(
+				'title' => 'Match Brackets',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'matchBrackets' , true),
+				'default' => false,
+				'set' => 'advanced',
+				'description' => 'highlight matching bracket pairs next to cursor',
+			),
+			'matchTags'=> array(
+				'title' => 'Match Tags',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'matchTags' , true),
+				'default' => false,
+				'set' => 'advanced',
+				'description' => 'highlight matching html tag pairs (not shortcodes)',
+			),
+			'highlightSelectionMatches'=> array(
+				'title' => 'Highlight Selection Matches',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'highlightSelectionMatches' , true),
+				'default' => false,
+				'set' => 'advanced',
+				'description' => 'highlight all instances of a currently selected word',
+			),
+			'autoCloseTags'=> array(
+				'title' => 'Auto Close Tags',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'autoCloseTags' , true),
+				'default' => false,
+				'set' => 'advanced',
+				'description' => 'auto-close HTML tags when <code>></code> or <code>/</code> is typed',
+			),
+			'autoCloseBrackets'=> array(
+				'title' => 'Auto Close Brackets',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'autoCloseBrackets' , true),
+				'default' => false,
+				'set' => 'advanced',
+				'description' => 'auto-close brackets and quotes when typed',
+			),
+			'foldGutter'=> array(
+				'title' => 'Code Folding',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'foldGutter' , true),
+				'default' => false,
+				'set' => 'advanced',
+				// 'description' => 'fold HTML code (not shortcodes)',
+			),
+			'scrollbarStyle'=> array(
+				'title' => 'Show Scrollbars',
+				'type' => 'checkbox',
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'scrollbarStyle' , true),
+				'default' => false,
+				'set' => 'advanced',
+			),
+			'keyMap'=> array(
+				'title' => 'Key Mapping',
+				'type' => 'radio',
+				'options' => array('default', 'sublime', 'vim', 'emacs'),
+				'current' => get_user_meta( get_current_user_id(), $this->prefix.'keyMap' , true),
+				'default' => 'default',
+				'set' => 'advanced',
 			),
 		);
-		// $this->addOns = array(
-		// 	'keyMap'=> array(
-		// 		'title' => 'Key Mapping',
-		//		'description' => 'choose a theme',
-		// 		'type' => 'select',
-		// 		'options' => array('none', 'emacs', 'sublime', 'vim'),
-		// 		'current' => get_user_meta( get_current_user_id(), $this->prefix.'keyMap' , true),
-		// 		'default' => 'none',
-		// 	),
-		// 	'styleActiveLine'=> array(),
-		// 	'matchBrackets'=> array(),
-		// 	'search'=> array(), // and replace
-		// 	'highlightSelectionMatches'=> array(),
-		// 	'styleSelectedText'=> array(), // ?
-		// 	'autoCloseBrackets'=> array(),
-		// 	'autoCloseTags'=> array(),
-		// 	'comment'=> array(), // continueComments?
-		// 	'foldCode'=> array(), // ?
-		// );
+
 
 		// Intalize all the hesh option fields as default if they don't exist yet
-		if (!get_user_meta( get_current_user_id(), $this->prefix.'hasInitalized', true) ) {
-			foreach ($this->userPrefrences as $id => $value) {
+		foreach ($this->userPrefrences as $id => $value) {
+			if (!isset($value['current']) || trim($value['current'])==='')
 				update_user_meta( get_current_user_id(), $this->prefix.$id, $value['default']);
-			}
-			update_user_meta( get_current_user_id(), $this->prefix.'hasInitalized', true);	
+			// error_log( $id . ': ' . $value['current'] . ': ' . isset($value['current']) ); // for debug
 		}
 	}
 
@@ -178,27 +227,26 @@ class wp_html_editor_syntax {
 		}
 	}
 
-	private function hesh_output_option($id, $config, $cm=false) {
+	private function hesh_output_option($id, $config) {
 		switch ($config['type']){
 			case 'select':
-				$this->hesh_output_select($id, $config, $cm);
+				$this->hesh_output_primary_select($id, $config);
 				break;
 			case 'checkbox':
-				$this->hesh_output_checkbox($id, $config, $cm);
+				$this->hesh_output_primary_toggle($id, $config);
 				break;
 		}
 	}
 
-	private function hesh_output_checkbox($id, $config, $cm=false) {
+	private function hesh_output_primary_toggle($id, $config) {
 		extract($config);
-		if ($cm): ?>
+		?>
 			<label 
 				class="CodeMirror-settings__button button button-small"
 				for="<?php echo $id; ?>"
-				<?php if (isset($description)) echo "title=\"$id-description\"" ?>
 				>
 				<?php echo $title; ?>
-				<input type="hidden" value="false"  name="<?php echo $id; ?>" />
+				<input type="hidden" value="false" name="<?php echo $id; ?>" />
 				<input 
 					name="<?php echo $id; ?>" 
 					id="<?php echo $id; ?>" 
@@ -206,51 +254,17 @@ class wp_html_editor_syntax {
 					value="true"
 					class="CodeMirror-settings__option"
 					<?php if ($current) echo 'checked'; ?>
-					/>
+				/>
 			</label>
-		<?php else: ?>
-			<tr>
-				<th scope="row">
-					<?php echo $title; ?>
-				</th>
-				<td>
-					<fieldset>
-						<legend class="screen-reader-text">
-							<span><?php echo $title; ?></span>
-						</legend>
-						<label for="<?php echo $id; ?>">
-							<input type="hidden" value="false"  name="<?php echo $id; ?>" />
-							<input 
-								name="<?php echo $id; ?>" 
-								id="<?php echo $id; ?>" 
-								type="checkbox"
-								value="true"
-								class="CodeMirror-settings__option"
-								<?php if (isset($description)) echo "aria-describedby=\"$id-description\"" ?>
-								<?php if ($current) echo 'checked'; ?>
-								/>
-							<?php echo $text; ?>
-						</label>
-						<?php if (isset($description)): ?>
-							<p class="description" 
-								id="<?php echo $id; ?>-description"
-								>
-								<?php echo $description; ?>
-							</p>
-						<?php endif; ?>
-					</fieldset>
-				</td>
-			</tr>
-		<?php endif;
+		<?php 
 	}
 	
-	private function hesh_output_select($id, $config, $cm=false) {
+	private function hesh_output_primary_select($id, $config) {
 		extract($config);
-		if ($cm): ?>
+		?>
 			<label 
 				class="CodeMirror-settings__button CodeMirror-settings__button--select button button-small"
 				for="<?php echo $id; ?>"
-				<?php if (isset($description)) echo "title=\"$id-description\"" ?>
 				>
 				<?php echo $title; ?>
 				<select 
@@ -268,38 +282,74 @@ class wp_html_editor_syntax {
 					<?php endforeach; ?>
 				</select>
 			</label>
-		<?php else: ?>
+		<?php
+	}
+
+	private function hesh_output_fieldset($title=false) {
+		if ($title): ?>
 			<tr>
 				<th scope="row">
-					<label for="<?php echo $id; ?>"><?php echo $title; ?></label>
+					<?php echo $title; ?>
 				</th>
-				<td>
-					<select 
-						id="<?php echo $id; ?>" 
-						name="<?php echo $id; ?>"
-						class="CodeMirror-settings__option"
-						<?php if (isset($description)) echo "aria-describedby=\"$id-description\"" ?>
-						>
-						<?php foreach ($options as $option): ?>
-							<option 
-								value="<?php echo $option; ?>"
-								<?php if ($current == $option) echo 'selected';?>
-								>
-								<?php echo ucfirst($option); ?>
-							</option>
-						<?php endforeach; ?>
-					</select>
-					<?php if (isset($description)): ?>
-						<p class="description" 
-							id="<?php echo $id; ?>-description"
-							>
-							<?php echo $description; ?>
-						</p>
-					<?php endif; ?>
-				</td>
+				<td><fieldset>
+		<?php else: ?>
+				</fieldset></td>
 			</tr>
 		<?php endif;
 	}
+
+	private function hesh_output_checkbox($id, $config) {
+		extract($config);
+		?>
+			<label 
+				for="<?php echo $id; ?>"
+				>
+				<input type="hidden" value="false" name="<?php echo $id; ?>" />
+				<input 
+					name="<?php echo $id; ?>" 
+					id="<?php echo $id; ?>" 
+					type="checkbox"
+					value="true"
+					class="CodeMirror-settings__option"
+					<?php if ($current) echo 'checked'; ?>
+				/>
+				<?php echo $title; ?>
+				<?php if (isset($description)): ?>
+					<small 
+					class="description CodeMirror-settings-advanced__description" 
+					id="<?php echo $id; ?>-description"
+					>
+					<?php echo $description; ?>
+				</small>
+				<?php endif; ?>
+			</label>
+			<br/>
+		<?php
+	}
+
+	private function hesh_output_radio($id, $config) {
+		extract($config);
+		$this->hesh_output_fieldset('Key Bindings'); 
+		foreach ($options as $option): 
+		?>
+			<label>
+				<input 
+					type="radio" 
+					id="<?php echo $id; ?>" 
+					name="<?php echo $id; ?>"
+					value="<?php echo $option; ?>"
+					class="CodeMirror-settings__option"
+					<?php if ($current == $option) echo 'checked'; ?>
+				/>
+				<span>
+					<?php echo ucfirst($option); ?>
+				</span>
+			</label><br/>
+		<?php 
+		endforeach; 
+		$this->hesh_output_fieldset(); 
+	}
+
 		
 	public function hesh_output_form() {
 		// ob_start();
@@ -314,33 +364,54 @@ class wp_html_editor_syntax {
 					<header class="CodeMirror-settings__header CodeMirror-settings__docked">
 						<?php
 							foreach ($this->userPrefrences as $id => $value) {
-								$this->hesh_output_option($id,$value,true);
+								if ( $value['set'] === 'advanced' ) continue;
+								$this->hesh_output_option($id,$value);
 							}
 						?>
-						<a class="CodeMirror-settings__toggle-advanced" 
+						<a 
+							class="CodeMirror-settings__toggle-advanced" 
 							id="CodeMirror-settings__toggle-advanced"
-							></a>
+						></a>
 					</header>
 					<div class="form CodeMirror-settings__body">
 						<?php wp_nonce_field($this->formProcessName,$this->nonceSecretCode);?>
 						<input name="action" value="hesh_options_form" type="hidden">
-						<?php $i=0; while ($i < 1): ?>
 						<table class="form-table"><tbody>
-							<tr><td class="CodeMirror-settings__heading"><h1>
-								Addons
-							</h1></td></tr>
-							<tr>
-								<th scope="row"><label>Coming Soon...</label></th>
-							</tr>
+							<tr><td class="CodeMirror-settings__heading" colspan="2">
+								<h1>
+									Advanced Options
+								</h1>
+								<!-- <small>These features are experimental and may have bugs</small> -->
+							</td></tr>
+							<?php
+
+							$this->hesh_output_fieldset('Highlighting'); 
+								$this->hesh_output_checkbox('matchBrackets',$this->userPrefrences['matchBrackets']); 
+								$this->hesh_output_checkbox('matchTags',$this->userPrefrences['matchTags']); 
+								$this->hesh_output_checkbox('highlightSelectionMatches',$this->userPrefrences['highlightSelectionMatches']); 
+							$this->hesh_output_fieldset(); 
+							
+							$this->hesh_output_fieldset('Auto Completion'); 
+								$this->hesh_output_checkbox('autoCloseTags',$this->userPrefrences['autoCloseTags']); 
+								$this->hesh_output_checkbox('autoCloseBrackets',$this->userPrefrences['autoCloseBrackets']); 
+							$this->hesh_output_fieldset(); 
+
+							$this->hesh_output_fieldset('Editor Tools'); 
+								$this->hesh_output_checkbox('foldGutter',$this->userPrefrences['foldGutter']); 
+								$this->hesh_output_checkbox('scrollbarStyle',$this->userPrefrences['scrollbarStyle']); 
+							$this->hesh_output_fieldset(); 
+
+							$this->hesh_output_radio('keyMap',$this->userPrefrences['keyMap']); 
+
+							?>
 						</tbody></table>
-						<?php $i++; endwhile; ?>
 					</div>
 					<footer class="CodeMirror-settings__footer CodeMirror-settings__docked">
 						<p class="CodeMirror-settings__foot-content CodeMirror-settings__feedback">
 							<small>Leave a 
 								<a href="https://wordpress.org/support/plugin/html-editor-syntax-highlighter/reviews/#new-post" target="_blank">review</a>,
 								fork on
-								<a href="https://github.com/mukhortov/HESH-WordPress-Plugin/issues/new" target="_blank">Github</a>, 
+								<a href="https://github.com/mukhortov/HESH-WordPress-Plugin/" target="_blank">Github</a>, 
 								or submit a 
 								<a href="https://github.com/mukhortov/HESH-WordPress-Plugin/issues/new" target="_blank">bug report or enhancement request</a>. 
 							</small>
@@ -360,6 +431,58 @@ class wp_html_editor_syntax {
 		<?php 
 		// return ob_get_clean();
 	}
+
+	private $cssThemes = array( 
+		'default',
+		'3024-day',
+		'3024-night',
+		'abcdef',
+		'ambiance-mobile',
+		'ambiance',
+		'base16-dark',
+		'base16-light',
+		'bespin',
+		'blackboard',
+		'cobalt',
+		'colorforth',
+		'dracula',
+		'duotone-dark',
+		'duotone-light',
+		'eclipse',
+		'elegant',
+		'erlang-dark',
+		'hopscotch',
+		'icecoder',
+		'isotope',
+		'lesser-dark',
+		'liquibyte',
+		'material',
+		'mbo',
+		'mdn-like',
+		'midnight',
+		'monokai',
+		'neat',
+		'neo',
+		'night',
+		'panda-syntax',
+		'paraiso-dark',
+		'paraiso-light',
+		'pastel-on-dark',
+		'railscasts',
+		'rubyblue',
+		'seti',
+		'solarized',
+		'the-matrix',
+		'tomorrow-night-bright',
+		'tomorrow-night-eighties',
+		'ttcn',
+		'twilight',
+		'vibrant-ink',
+		'xq-dark',
+		'xq-light',
+		'yeti',
+		'zenburn'
+	);		  
 
 }